HTTP GET, PUT and Delete for Order and Notification

+++ Notification/src/main/java/br/edu/anhembi/model/repository/NotificationRepository.java
@@ -3,5 +3,5 @@ package br.edu.anhembi.model.repository;
 import br.edu.anhembi.model.Notifications;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notifications, Integer> {
+public interface NotificationRepository extends JpaRepository<Notifications, Long> {
 }
